tools: graphpkgfetcher: refactor as a lib instead of main package
